# Conflicts:
#	src/main/java/com/jean/config/context/AppConfig.java
#	src/main/java/com/jean/config/security/jwt/JwtTokenUtil.java
#	src/main/java/com/jean/dao/UserDao.java
#	src/main/java/com/jean/dao/entity/User.java
#	src/main/java/com/jean/dao/impl/UserDaoImpl.java
#	src/main/java/com/jean/service/impl/UserServiceImpl.java
#	src/main/java/com/jean/servlet/controller/UserRestController.java

+++ src/main/java/com/jean/BehaviorConstants.java
@@ -5,34 +5,34 @@ package com.jean;
  */
 public interface BehaviorConstants {
 
-	double CRITICAL_HIGH_TEMP = 28.5;
+	float CRITICAL_HIGH_TEMP = 28.5f;
 
-	double MIN_PRESSURE_CHANGES = 1;
-	double MAX_PRESSURE_CHANGES = 2.5;
+	float MIN_PRESSURE_CHANGES = 1;
+	float MAX_PRESSURE_CHANGES = 2.5f;
 
-	double CRITICAL_WIND_SPEED = 6;
-	double MAX_WIND_SPEED = 3.3;
+	float CRITICAL_WIND_SPEED = 6;
+	float MAX_WIND_SPEED = 3.3f;
 
-	double HIGH_PRESSURE_LEVEL = 759.0;
-	double LOW_PRESSURE_LEVEL = 744.00;
+	float HIGH_PRESSURE_LEVEL = 759.0f;
+	float LOW_PRESSURE_LEVEL = 744.00f;
 
-	double MAX_RAIN_VOLUME = 3.3;
-	double MIN_RAIN_VOLUME = 0.3;
+	float MAX_RAIN_VOLUME = 3.3f;
+	float MIN_RAIN_VOLUME = 0.3f;
 
-	double CRASH_NIBBLE_POINT = -15;
-	double SPAWNING_POINT = -10;
-	double COUNT_OF_POINT = 10;
-	double HIGH_PRESSURE_POINT = 8;
-	double LOW_PRESSURE_POINT = 6;
-	double RISE_PRESSURE_POINT = 6;
-	double DOWN_PRESSURE_POINT = 10;
-	double STABILITY_PRESSURE_POINT = 8;
-	double UNSTABILITY_PRESSURE_POINT = 2;
+	float CRASH_NIBBLE_POINT = -15;
+	float SPAWNING_POINT = -10;
+	float COUNT_OF_POINT = 10;
+	float HIGH_PRESSURE_POINT = 8;
+	float LOW_PRESSURE_POINT = 6;
+	float RISE_PRESSURE_POINT = 6;
+	float DOWN_PRESSURE_POINT = 10;
+	float STABILITY_PRESSURE_POINT = 8;
+	float UNSTABILITY_PRESSURE_POINT = 2;
 
-	double LIGHT_RAIN_POINT = -1;
-	double HEAVY_RAIN_POINT = -5;
-	double WIND_POINT = -3;
-	double DAY_ACTIVITY_POINT = 1;
+	float LIGHT_RAIN_POINT = -1;
+	float HEAVY_RAIN_POINT = -5;
+	float WIND_POINT = -3;
+	float DAY_ACTIVITY_POINT = 1;
 	
 	String SPAWNING_MESSAGE = "Is it spawning now!";
