+++ src/main/java/com/jean/service/BaitService.java
@@ -3,11 +3,11 @@ package com.jean.service;
 import java.util.List;
 
 import com.jean.DaoDfmException;
-import com.jean.entity.bait.Bait;
+import com.jean.dao.entity.bait.Bait;
 
 public interface BaitService {
 
-	List<Bait> getBaits(Integer baitId, String baitType, String baitName, Integer fishId, String date) throws DaoDfmException;
+	List<Bait> getBaits(Integer baitId, String baitType, String baitName, Long fishId, String date) throws DaoDfmException;
 	
 	Integer saveBait(Bait bait);
